check if backup file exists and skip if it does

+++ misc/vscodestatusbarcolor.sh
@@ -28,10 +28,14 @@ fi
 if [ ! -d ~/.vscode ];then
     mkdir ~/.vscode
 fi
-cp "$DIR"/resources/app/out/vs/workbench/electron-browser/workbench.main.css ~/.vscode/workbench.main.css.backup
+if [ ! -f ~/.vscode/workbench.main.css.backup ];then
+    cp "$DIR"/resources/app/out/vs/workbench/electron-browser/workbench.main.css ~/.vscode/workbench.main.css.backup
+    echo "Backup workbench.main.css.backup created in '~/.vscode/'; use '--revert' to restore it."
+else
+    echo "workbench.main.css.backup exists; skipping backup; use '--revert' to restore it."
+fi
 sudo sed -i -e 's/007acc/'$REPLY'/g' "$DIR"/resources/app/out/vs/workbench/electron-browser/workbench.main.css
 echo "Status bar color changed to #$REPLY!"
-echo "Backup workbench.main.css created in '~/.vscode/'; use '--revert' to restore it."
 echo "Trying to change your color a second time without first using '--revert' will have no effect."
 echo "Restarting 'VSCode'..."
 /usr/bin/code
